requirements.txt
asgiref==3.7.2
Django==4.2.7
django-crontab==0.7.1
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==20.1.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
